fix(knobs): declare CORE_UTILIZATION as int so sampler/log/variant/emission agree
CORE_UTILIZATION was type="float" (range 20.0-60.0) but the funnel truncates
it to int before emitting the ORFS build (int(...) at F3), and ORFS treats it
as an integer percentage. The result (audit F9): the Optuna study and campaign
logs recorded ~1125 distinct sub-integer values on likith that the build never
saw (its effective axis is ~5 levels, 3..7), TPE fit structure to pure noise,
and two "different" configs (6.01 vs 6.99) produced the same build yet hashed to
different variant dirs -> cache misses and duplicate physical builds.

Declaring it type="int" (range 20-60, default 40) makes the sampled value, the
log, the variant hash, and the emission agree. Per-design float override ranges
(likith [3.0, 7.0], sagar [10.0, 20.0]) still work unchanged: space() feeds them
through suggest_int(int(lo), int(hi)). Belt-and-braces, the funnel emission is
now int(round(float(...))) so a design-fixed float (e.g. default 5.0) rounds
rather than truncates (6.99 -> 7, not 6).

Done by an AI agent

+++ eda_rl/common/knobs.py
@@ -557,9 +557,16 @@ def _build_knobs() -> list[Knob]:
             name="CORE_UTILIZATION",
             tier=1,
             stage="floorplan",
-            type="float",
-            range=(20.0, 60.0),
-            default=40.0,
+            # int, not float: ORFS emits CORE_UTILIZATION as an integer percentage
+            # and the funnel truncates it to int before the build, so a sampled
+            # 6.01 vs 6.99 are the SAME physical build but hashed to different
+            # variant dirs (cache misses) and shown to Optuna as distinct levels
+            # (audit F9).  Declaring it int makes sampler/log/variant/emission
+            # agree.  Per-design float override ranges (e.g. likith [3.0, 7.0])
+            # still work: space() runs suggest_int(int(lo), int(hi)).
+            type="int",
+            range=(20, 60),
+            default=40,
             _emit_template="export CORE_UTILIZATION = {v}",
             notes=(
                 "Verified tunable=1 in variables.yaml (stages: floorplan). "
